Styles for all layers
It is far from complete, but it contains many attributes form settings that make editing objects easier.
For example, the names of mandatory fields are red, optional mandatory fields (to be filled in by the land registry office or the system) are green, and optional fields are black.
OBJ_FELS, RETEG_ID, TELEPULES_ID, FEKVES, HRSZ values can be selected from a list
etc, etc